refactor(auth): split Register into view + hook modules
Register.tsx was 347 lines mixing form state + validation + async
handlers with three different full-page views (duplicate-email,
success, and the form itself). Split them so the route picks a view
and everything stateful lives in one hook:

- `register/useRegister.ts` — form state, per-field validation,
  server/duplicate/success flags, loading flags, handleChange,
  handleSubmit (with registerSchema + DUPLICATE_EMAIL mapping),
  handleGoogleSignUp.
- `register/DuplicateEmailView.tsx` — terminal screen with the three
  recovery CTAs (login, reset, support).
- `register/SuccessView.tsx` — post-register confirmation; teacher
  variant surfaces the admin-approval banner.
- `register/RegisterForm.tsx` — role selector, four inputs, Google
  OAuth shortcut. Purely controlled by props.
- `register/GoogleIcon.tsx` — the inline brand SVG (extracted from
  the page body).
- `Register.tsx` — 42-line thin orchestrator: reads `useRegister`
  and picks one of the three views.

No behaviour change. All error messages, validation rules, and
terminal-state flow are preserved 1:1.

Verified locally: tsc, eslint, knip ({"issues":[]}), vitest (85
tests) all pass.

Made-with: Cursor
